Move reusable components to subfolder
