+++ src/components/__tests__/reusables/SaveButton-test.tsx
@@ -4,7 +4,7 @@ import * as Enzyme from "enzyme";
 import * as React from "react";
 import { shallow, mount } from "enzyme";
 
-import SaveButton from "../SaveButton";
+import SaveButton from "../../reusables/SaveButton";
 
 describe("SaveButton", () => {
   let wrapper: Enzyme.ShallowWrapper<any, {}>;
