+++ src/stylesheets/app.scss
@@ -3,10 +3,10 @@
 
 @import "mixins";
 @import "colors";
-@import "form";
+@import "reusables/form";
 @import "global";
 @import "icon";
 @import "list";
-@import "panel";
+@import "reusables/panel";
 @import "header";
 @import "detail_page";
