+++ src/stylesheets/reusables/panel.scss
@@ -9,7 +9,6 @@
   justify-content: space-between;
   align-items: center;
   transition: background .5s;
-
   span {
     color: $black;
     transition: color .5s;
